+++ README.md
@@ -1,2 +1,17 @@
 # BibsClient
 UI Controls for BibsLeague app
+
+
+## Firebase
+
+```
+firebase login
+```
+
+```
+firebase init
+```
+
+```
+firebase deploy
+```
