+++ src/bibs-portal/firebase.json
@@ -1,13 +1,16 @@
 {
-    "hosting": {
-        "public": "dist/bibs-portal",
-        "ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
-        "redirects": [],
-        "rewrites": [
-            {
-                "source": "**",
-                "destination": "/index.html"
-            }
-        ]
-    }
+  "hosting": {
+    "public": "public",
+    "ignore": [
+      "firebase.json",
+      "**/.*",
+      "**/node_modules/**"
+    ],
+    "rewrites": [
+      {
+        "source": "**",
+        "destination": "/index.html"
+      }
+    ]
+  }
 }
