+++ src/bibs-portal/src/app/components/players/player-list-item/player-list-item.component.scss
