+++ src/bibs-portal/src/app/components/players/player-list/player-list.component.scss
