+++ src/bibs-portal/src/app/core/players-table/players-table.component.html
@@ -25,7 +25,7 @@
         <th mat-header-cell *matHeaderCellDef> Actions </th>   
         <td mat-cell *matCellDef="let element"> 
           <!-- <button mat-raised-button >Edit</button>  -->
-          <button mat-raised-button color="primary" (click)="onEdit(element)">Edit</button>
+          <button mat-raised-button id="editbtn" color="primary" (click)="onEdit(element)">Edit</button>
           <button mat-raised-button color="warn" (click)="onDelete(element)">Delete</button>
         </td>
       </ng-container>
