+++ src/bibs-portal/src/app/core/players-table/players-table.component.scss
@@ -6,3 +6,7 @@ table {
   display: flex;
   justify-content: center;
 }
+
+#editbtn {
+  margin-right: 15px;
+}
